refactor(backup-archive): move backup disk-read for import
Move createVhdDisk to the disks module, readFullVmBackup to VmFullBackupArchive and
readIncrementalVmBackup to VmIncrementalBackupArchive, all taking handler explicitly.
RemoteAdapter delegates to them; the ancestor-chain opening and ReadAhead wrapping
now live next to the other disk helpers.
