+++ @xen-orchestra/backup-archive/src/index.mts
@@ -1,2 +1,4 @@
 export { VmBackupDirectory } from './VmBackupDirectory.mjs'
+export { VmFullBackupArchive } from './VmFullBackupArchive.mjs'
+export { VmIncrementalBackupArchive } from './VmIncrementalBackupArchive.mjs'
 export { deleteMetadataBackup, deleteOldMetadataBackups } from './MetadataBackupDirectory.mjs'
